2025-12-26-stanford-storm-framework/action-in-blog/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.3
boto3==1.42.17
botocore==1.42.17
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
courlan==1.3.2
datasets==2.14.7
dateparser==1.2.2
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dspy-ai==2.4.9
faiss-cpu==1.13.2
Faker==39.0.0
fastuuid==0.14.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2023.10.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jusText==3.0.2
knowledge-storm==1.1.1
langchain-core==1.2.5
langchain-huggingface==1.2.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langsmith==0.5.1
litellm==1.80.0
lxml==6.0.2
lxml_html_clean==0.4.3
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
networkx==3.6.1
numpy==2.4.0
openai==1.109.1
optuna==4.6.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pyarrow-hotfix==0.7
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
ruff==0.14.10
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tld==0.13.1
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
uuid_utils==0.12.0
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
